DecoupleR network format

A network passed to any run_ method in the package must contain at least two attributes: .source and .target. In addition, the methods must map their corresponding metadata associated with their edges.

Usage,NULL

Arguments

network

Tibble or dataframe with edges and it's associated metadata.

.source

Column with source nodes.

.target

Column with target nodes.

.mor

Column with edge mode of regulation (i.e. mor).

.likelihood

Column with edge likelihood.

Details

All the attributes to be mapped are prefixed by .

The idea of using this type of mapping is to provide flexibility to different types of networks, be they regulatory, metabolic, or of any other type. This way, you should only consider having your network or networks in a long format and these can easily be manipulated by functions within the tidyverse ecosystem.

See also