Pre-processing for methods that fit models

Get only the intersection of target features between mat and network.

Transform tidy network into matrix representation with mor as value.

If center is true, then the expression values are centered by the mean of expression across the samples.

.fit_preprocessing(network, mat, center, na.rm, sparse)

Arguments

network

Tibble or dataframe with edges and it's associated metadata.

mat

Matrix to evaluate (e.g. expression matrix). Target nodes in rows and conditions in columns. rownames(mat) must have at least one intersection with the elements in network .target column.

center

Logical value indicating if mat must be centered by base::rowMeans().

na.rm

Should missing values (including NaN) be omitted from the calculations of base::rowMeans()?

sparse

Deprecated parameter.

Value

A named list of matrices to evaluate in methods that fit models, like .mlm_analysis().

mat: Features as rows and samples as columns.

mor_mat: Features as rows and columns as source.