Wrapper to access resources inside Omnipath. This wrapper allows to easily query different prior knowledge resources. To check available resources run decoupleR::show_resources(). For more information visit the official website for Omnipath.

Wrapper to access resources inside Omnipath. This wrapper allows to easily query different prior knowledge resources. To check available resources run decoupleR::show_resources(). For more information visit the official website for Omnipath.

Usage

get_resource(name, organism = "human", ...)

Arguments

name

Name of the resource to query.

organism

Organism name or NCBI Taxonomy ID.

...

Passed to OmnipathR::import_omnipath_annotations.

Examples

df <- decoupleR::get_resource('SIGNOR')