Intersect network target genes with expression matrix.

Keep only edges which its target genes belong to the expression matrix.

intersect_regulons(mat, network, .source, .target, minsize)

Arguments

mat
Matrix to evaluate (e.g. expression matrix). Target nodes in rows and conditions in columns. rownames(mat) must have at least one intersection with the elements in network .target column.

network
Tibble or dataframe with edges and it's associated metadata.

.source
Column with source nodes.

.target
Column with target nodes.

minsize
Minimum number of targets per source allowed.

Value

Filtered tibble.

Examples

inputs_dir <- system.file("testdata", "inputs", package = "decoupleR")
mat <- readRDS(file.path(inputs_dir, "input-expr_matrix.rds"))
network <- readRDS(file.path(inputs_dir, "input-dorothea_genesets.rds"))
intersect_regulons(mat, network, tf, target, minsize=5)

#> # A tibble: 150 × 5
#> # Groups:   tf [5]
#>    tf    confidence target     mor likelihood
#>    <chr> <chr>      <chr>    <dbl>      <dbl>
#>  1 FOXO4 A          BCL2L11      1          1
#>  2 FOXO4 A          BCL6         1          1
#>  3 FOXO4 A          CDKN1A       1          1
#>  4 FOXO4 A          CDKN1B       1          1
#>  5 FOXO4 A          G6PC         1          1
#>  6 FOXO4 A          GADD45A      1          1
#>  7 FOXO4 A          IGFBP1       1          1
#>  8 FOXO4 A          PCK2         1          1
#>  9 FOXO4 A          PPARGC1A     1          1
#> 10 FOXO4 A          RBL2         1          1
#> # … with 140 more rows