AUCell

Calculates regulatory activities using Area Under the Curve (AUC) from AUCell

run_aucell(
  mat,
  network,
  .source = .data$source,
  .target = .data$target,
  aucMaxRank = ceiling(0.05 * nrow(rankings)),
  nproc = 4,
  seed = 42
)

Arguments

mat
Matrix to evaluate (e.g. expression matrix). Target nodes in rows and conditions in columns. rownames(mat) must have at least one intersection with the elements in network .target column.

network
Tibble or dataframe with edges and it's associated metadata.

.source
Column with source nodes.

.target
Column with target nodes.

aucMaxRank
Threshold to calculate the AUC.

nproc
Number of cores to use for computation.

seed
A single value, interpreted as an integer, or NULL for random number generation.

Details

This function is a wrapper for the method AUCell. It uses the "Area Under the Curve" (AUC) to calculate whether a critical subset of input molecular features is enriched for each sample.

See also

Examples

inputs_dir <- system.file("testdata", "inputs", package = "decoupleR")

mat <- readRDS(file.path(inputs_dir, "input-expr_matrix.rds"))
network <- readRDS(file.path(inputs_dir, "input-dorothea_genesets.rds"))

run_aucell(mat, network, .source='tf', nproc=1)

#> # A tibble: 20 × 4
#>    statistic source condition   score
#>    <chr>     <chr>  <chr>       <dbl>
#>  1 aucell    FOXO4  GSM2753335 0.0566
#>  2 aucell    FOXO4  GSM2753336 0.0613
#>  3 aucell    FOXO4  GSM2753337 0.0985
#>  4 aucell    FOXO4  GSM2753338 0.120 
#>  5 aucell    NFIC   GSM2753335 0.0313
#>  6 aucell    NFIC   GSM2753336 0.0339
#>  7 aucell    NFIC   GSM2753337 0.0418
#>  8 aucell    NFIC   GSM2753338 0.0708
#>  9 aucell    RFXAP  GSM2753335 0     
#> 10 aucell    RFXAP  GSM2753336 0.0437
#> 11 aucell    RFXAP  GSM2753337 0.114 
#> 12 aucell    RFXAP  GSM2753338 0.126 
#> 13 aucell    SMAD3  GSM2753335 0.0405
#> 14 aucell    SMAD3  GSM2753336 0.0442
#> 15 aucell    SMAD3  GSM2753337 0.0722
#> 16 aucell    SMAD3  GSM2753338 0.0893
#> 17 aucell    TFAP2A GSM2753335 0.0161
#> 18 aucell    TFAP2A GSM2753336 0.0155
#> 19 aucell    TFAP2A GSM2753337 0.0304
#> 20 aucell    TFAP2A GSM2753338 0.0557