Multivariate Linear Model (MLM)

Calculates regulatory activities by fitting multivariate linear models (MLM)

Usage

run_mlm(
  mat,
  network,
  .source = .data$source,
  .target = .data$target,
  .mor = .data$mor,
  .likelihood = .data$likelihood,
  sparse = FALSE,
  center = FALSE,
  na.rm = FALSE
)

Arguments

mat

Matrix to evaluate (e.g. expression matrix). Target nodes in rows and conditions in columns. rownames(mat) must have at least one intersection with the elements in network .target column.

network

Tibble or dataframe with edges and it's associated metadata.

.source

Column with source nodes.

.target

Column with target nodes.

.mor

Column with edge mode of regulation (i.e. mor).

.likelihood

Column with edge likelihood.

sparse

Deprecated parameter.

center

Logical value indicating if mat must be centered by base::rowMeans().

na.rm

Should missing values (including NaN) be omitted from the calculations of base::rowMeans()?

Value

A long format tibble of the enrichment scores for each source across the samples. Resulting tibble contains the following columns:

statistic: Indicates which method is associated with which score.

source: Source nodes of network.

condition: Condition representing each column of mat.

score: Regulatory activity (enrichment score).

Details

MLM fits a multivariate linear model to estimate regulatory activities. MLM transforms a given network into an adjacency matrix, placing sources as columns and targets as rows. The matrix is filled with the associated weights for each interaction. This matrix is used to fit a linear model to predict the observed molecular readouts per sample. The obtained t-values from the fitted model are the activities of the regulators.

See also

Examples

inputs_dir <- system.file("testdata", "inputs", package = "decoupleR")

mat <- readRDS(file.path(inputs_dir, "input-expr_matrix.rds"))
network <- readRDS(file.path(inputs_dir, "input-dorothea_genesets.rds"))

run_mlm(mat, network, .source='tf')
#> # A tibble: 20 × 5
#>    statistic source condition  score p_value
#>    <chr>     <chr>  <chr>      <dbl>   <dbl>
#>  1 mlm       FOXO4  GSM2753335 2.21  0.0288 
#>  2 mlm       NFIC   GSM2753335 1.12  0.263  
#>  3 mlm       SMAD3  GSM2753335 0.696 0.488  
#>  4 mlm       TFAP2A GSM2753335 1.34  0.183  
#>  5 mlm       RFXAP  GSM2753335 1.63  0.105  
#>  6 mlm       FOXO4  GSM2753336 2.16  0.0325 
#>  7 mlm       NFIC   GSM2753336 1.16  0.250  
#>  8 mlm       SMAD3  GSM2753336 0.882 0.380  
#>  9 mlm       TFAP2A GSM2753336 1.56  0.122  
#> 10 mlm       RFXAP  GSM2753336 2.31  0.0226 
#> 11 mlm       FOXO4  GSM2753337 2.37  0.0195 
#> 12 mlm       NFIC   GSM2753337 0.729 0.467  
#> 13 mlm       SMAD3  GSM2753337 1.11  0.270  
#> 14 mlm       TFAP2A GSM2753337 1.54  0.126  
#> 15 mlm       RFXAP  GSM2753337 2.85  0.00507
#> 16 mlm       FOXO4  GSM2753338 2.10  0.0378 
#> 17 mlm       NFIC   GSM2753338 0.550 0.584  
#> 18 mlm       SMAD3  GSM2753338 0.860 0.391  
#> 19 mlm       TFAP2A GSM2753338 1.28  0.204  
#> 20 mlm       RFXAP  GSM2753338 2.72  0.00742