PSCIRA (Permutation Single Cell Inference of Regulatory Activity)

Calculate the regulatory activity of each tf by multiplying the expression values of its objectives with their corresponding associated profiles for each given condition.The result is equal to the z-score of the found value compared to its null distribution.

run_pscira(
  mat,
  network,
  .source = .data$tf,
  .target = .data$target,
  .mor = .data$mor,
  sparse = TRUE,
  times = 10,
  seed = 42
)

Arguments

mat
Matrix to evaluate (e.g. expression matrix). Target nodes in rows and conditions in columns. rownames(mat) must have at least one intersection with the elements in network .target column.

network
Tibble or dataframe with edges and it's associated metadata.

.source
Column with source nodes.

.target
Column with target nodes.

.mor
Column with edge mode of regulation (i.e. mor).

sparse
Logical value indicating if the generated profile matrix should be sparse.

times
Number of replications.

seed
A single value, interpreted as an integer, or NULL.

Value

A long format tibble of the enrichment scores for each tf across the conditions. Resulting tibble contains the following columns:

statistic: Indicates which method is associated with which score.

tf: Source nodes of network.

condition: Condition representing each column of mat.

score: Regulatory activity (enrichment score).

statistic_time: Internal execution time indicator.

Details

Estimation of regulatory activity: A linear regression of the expression profile is performed against the "target profile" of the given TF, where in the target profile, any member of the regulon will be assigned a +1 for activated actions, a -1 for inhibitory activations and, finally, a 0 for all genes that are not a member of the TF regulon. TF activity is then defined as the t-statistic of this linear regression.

See also

Examples

inputs_dir <- system.file("testdata", "inputs", package = "decoupleR")

mat <- readRDS(file.path(inputs_dir, "input-expr_matrix.rds"))
network <- readRDS(file.path(inputs_dir, "input-dorothea_genesets.rds"))

run_pscira(mat, network, tf, target, mor)

#> # A tibble: 20 x 5
#>    statistic tf     condition  score statistic_time
#>    <chr>     <chr>  <chr>      <dbl> <drtn>        
#>  1 pscira    FOXO4  GSM2753335 2.50  0.5356321 secs
#>  2 pscira    FOXO4  GSM2753336 2.57  0.5356321 secs
#>  3 pscira    FOXO4  GSM2753337 3.05  0.5356321 secs
#>  4 pscira    FOXO4  GSM2753338 2.80  0.5356321 secs
#>  5 pscira    NFIC   GSM2753335 0.711 0.5356321 secs
#>  6 pscira    NFIC   GSM2753336 0.791 0.5356321 secs
#>  7 pscira    NFIC   GSM2753337 0.584 0.5356321 secs
#>  8 pscira    NFIC   GSM2753338 0.650 0.5356321 secs
#>  9 pscira    RFXAP  GSM2753335 1.19  0.5356321 secs
#> 10 pscira    RFXAP  GSM2753336 1.79  0.5356321 secs
#> 11 pscira    RFXAP  GSM2753337 2.42  0.5356321 secs
#> 12 pscira    RFXAP  GSM2753338 2.44  0.5356321 secs
#> 13 pscira    SMAD3  GSM2753335 0.353 0.5356321 secs
#> 14 pscira    SMAD3  GSM2753336 0.557 0.5356321 secs
#> 15 pscira    SMAD3  GSM2753337 1.04  0.5356321 secs
#> 16 pscira    SMAD3  GSM2753338 1.08  0.5356321 secs
#> 17 pscira    TFAP2A GSM2753335 1.42  0.5356321 secs
#> 18 pscira    TFAP2A GSM2753336 1.58  0.5356321 secs
#> 19 pscira    TFAP2A GSM2753337 1.88  0.5356321 secs
#> 20 pscira    TFAP2A GSM2753338 2.07  0.5356321 secs