SCIRA (Single Cell Inference of Regulatory Activity)

Calculates TF activity according to SCIRA.

run_scira(
  emat,
  regulons,
  .source = .data$tf,
  .target = .data$target,
  .target_profile = .data$mor,
  .sparse = FALSE
)

Arguments

emat
A named expression matrix. E.g Genes in rows and samples in columns.

regulons
A data frame of regulons in table format.

.source
Column name in regulons with the factors to calculate the enrichment scores.

.target
Column name in regulons that relates the labels of the rows in the expression matrix.

.target_profile
Column name in regulons that contains the target profile. E.g. MoR.

.sparse
Logical value indicating if the generated profile matrix should be sparse.

Value

A long format tibble of the enrichment results for each set of genes across the samples. Resulting tibble contains the following columns:

source Resources from the .source column of the regulons data frame.

condition Samples representing each column of emat.

score Regulatory activity of each resource for each condition.

Details

Estimation of regulatory activity: A linear regression of the expression profile is performed against the "target profile" of the given TF, where in the target profile, any member of the regulon will be assigned a +1 for activated actions, a -1 for inhibitory activations and, finally, a 0 for all genes that are not a member of the TF regulon. TF activity is then defined as the t-statistic of this linear regression.