Univariate Decision Tree (UDT)

Calculates regulatory activities by fitting univariate decision trees (UDT) using rpart::rpart().

Usage,

run_udt(
  mat,
  network,
  .source = .data$source,
  .target = .data$target,
  .mor = .data$mor,
  .likelihood = .data$likelihood,
  sparse = FALSE,
  center = FALSE,
  na.rm = FALSE,
  min_n = 20,
  seed = 42
)

Arguments

mat

Matrix to evaluate (e.g. expression matrix). Target nodes in rows and conditions in columns. rownames(mat) must have at least one intersection with the elements in network .target column.

network

Tibble or dataframe with edges and it's associated metadata.

.source

Column with source nodes.

.target

Column with target nodes.

.mor

Column with edge mode of regulation (i.e. mor).

.likelihood

Column with edge likelihood.

sparse

Deprecated parameter.

center

Logical value indicating if mat must be centered by base::rowMeans().

na.rm

Should missing values (including NaN) be omitted from the calculations of base::rowMeans()?

min_n

An integer for the minimum number of data points in a node that are required for the node to be split further.

seed

A single value, interpreted as an integer, or NULL for random number generation.

Value

A long format tibble of the enrichment scores for each source across the samples. Resulting tibble contains the following columns:

statistic: Indicates which method is associated with which score.

source: Source nodes of network.

condition: Condition representing each column of mat.

score: Regulatory activity (enrichment score).

Details

UDT fits a (univariate) decision tree to estimate regulatory activities. UDT fits a decision tree that predicts the observed molecular readouts using the given weights of a regulator as a single co-variate. The obtained feature importance from the fitted model is the activity of the regulator.

See also

Examples

inputs_dir <- system.file("testdata", "inputs", package = "decoupleR")

mat <- readRDS(file.path(inputs_dir, "input-expr_matrix.rds"))
network <- readRDS(file.path(inputs_dir, "input-dorothea_genesets.rds"))

run_udt(mat, network, .source='tf')
#> # A tibble: 20 × 4
#>    statistic source condition  score
#>    <chr>     <chr>  <chr>      <dbl>
#>  1 udt       FOXO4  GSM2753335  1.78
#>  2 udt       FOXO4  GSM2753336  1.48
#>  3 udt       FOXO4  GSM2753337  2.87
#>  4 udt       FOXO4  GSM2753338  2.21
#>  5 udt       NFIC   GSM2753335  4.42
#>  6 udt       NFIC   GSM2753336  4.21
#>  7 udt       NFIC   GSM2753337  4.57
#>  8 udt       NFIC   GSM2753338  2.38
#>  9 udt       SMAD3  GSM2753335  5.54
#> 10 udt       SMAD3  GSM2753336  3.21
#> 11 udt       SMAD3  GSM2753337  4.26
#> 12 udt       SMAD3  GSM2753338  4.65
#> 13 udt       TFAP2A GSM2753335  5.41
#> 14 udt       TFAP2A GSM2753336  8.34
#> 15 udt       TFAP2A GSM2753337  8.56
#> 16 udt       TFAP2A GSM2753338  9.49
#> 17 udt       RFXAP  GSM2753335  2.08
#> 18 udt       RFXAP  GSM2753336  1.58
#> 19 udt       RFXAP  GSM2753337  3.21
#> 20 udt       RFXAP  GSM2753338  2.28