Weighted Mean (WMEAN)

Calculates regulatory activities by computing the WMEAN.

Usage,

run_wmean(
  mat,
  network,
  .source = .data$source,
  .target = .data$target,
  .mor = .data$mor,
  .likelihood = .data$likelihood,
  times = 100,
  seed = 42,
  sparse = TRUE,
  randomize_type = "rows"
)

Arguments

mat

Matrix to evaluate (e.g. expression matrix). Target nodes in rows and conditions in columns. rownames(mat) must have at least one intersection with the elements in network .target column.

network

Tibble or dataframe with edges and it's associated metadata.

.source

Column with source nodes.

.target

Column with target nodes.

.mor

Column with edge mode of regulation (i.e. mor).

.likelihood

Column with edge likelihood.

times

How many permutations to do?

seed

A single value, interpreted as an integer, or NULL for random number generation.

sparse

Should the matrices used for the calculation be sparse?

randomize_type

How to randomize the expression matrix.

Value

A long format tibble of the enrichment scores for each source across the samples. Resulting tibble contains the following columns:

statistic: Indicates which method is associated with which score.

source: Source nodes of network.

condition: Condition representing each column of mat.

score: Regulatory activity (enrichment score).

p_value: p-value for the score of the method.

Details

Infers activity score for each regulator by weighting the molecular readouts of its targets by their mode of regulations and likelihoods. In addition, it runs permutations to calculate empirical p-values, providing normalized (z-score) and corrected activity (estimate * -log10(pval)) scores. This is represented in the statistic column which will contain three values for each call to run_wmean(); wmean, norm_wmean and corr_wmean.

See also

Examples

inputs_dir <- system.file("testdata", "inputs", package = "decoupleR")

mat <- readRDS(file.path(inputs_dir, "input-expr_matrix.rds"))
network <- readRDS(file.path(inputs_dir, "input-dorothea_genesets.rds"))

run_wmean(mat, network, .source='tf')
#> # A tibble: 60 × 5
#>    statistic  source condition  score p_value
#>    <chr>      <chr>  <chr>      <dbl>   <dbl>
#>  1 corr_wmean FOXO4  GSM2753335  9.13    0.01
#>  2 corr_wmean FOXO4  GSM2753336  9.18    0.01
#>  3 corr_wmean FOXO4  GSM2753337  9.77    0.01
#>  4 corr_wmean FOXO4  GSM2753338  9.84    0.01
#>  5 corr_wmean NFIC   GSM2753335  1.58    0.23
#>  6 corr_wmean NFIC   GSM2753336  1.84    0.2 
#>  7 corr_wmean NFIC   GSM2753337  1.23    0.27
#>  8 corr_wmean NFIC   GSM2753338  1.27    0.28
#>  9 corr_wmean RFXAP  GSM2753335  3.61    0.11
#> 10 corr_wmean RFXAP  GSM2753336  6.51    0.04
#> # … with 50 more rows