Convert a network to run under the method of interest.

Convert a long-format network to the suggested standard for the specified run_{statistic}(). If the default parameters are not modified, then the function sets its own null values for those columns.

convert_to_(network)

convert_to_aucell(network, .source, .target)

convert_to_ulm(network, .source, .target, .mor = NULL, .likelihood = NULL)

convert_to_mlm(network, .source, .target, .mor = NULL, .likelihood = NULL)

convert_to_wsum(network, .source, .target, .mor = NULL, .likelihood = NULL)

convert_to_wmean(network, .source, .target, .mor = NULL, .likelihood = NULL)

convert_to_viper(network, .source, .target, .mor = NULL, .likelihood = NULL)

convert_to_gsva(network, .source, .target)

convert_to_ora(network, .source, .target)

convert_to_fgsea(network, .source, .target)

Arguments

network
Tibble or dataframe with edges and it's associated metadata.

.source
Column with source nodes.

.target
Column with target nodes.

.mor
Column with edge mode of regulation (i.e. mor).

.likelihood
Column with edge likelihood.

Value

convert_to_ Return same as input.

convert_to_aucell() Return a named list of sources with associated targets.

convert_to_gsva() Return a list of sources with associated targets suitable for GSVA::gsva().

convert_to_wmean() Return a tibble with four columns: source, target, mor and likelihood.

convert_to_ora() Return a named list of sources with associated targets.

convert_to_wsum() Returns a tibble with three columns: source, target and mor.

convert_to_ulm() Returns a tibble with three columns: source, target and mor.

convert_to_mlm() Returns a tibble with three columns: source, target and mor.

convert_to_viper() Return a list of sources with associated targets suitable for viper::viper()

See also

Examples

inputs_dir <- system.file("testdata", "inputs", package = "decoupleR")

network <- readRDS(file.path(inputs_dir, "input-dorothea_genesets.rds"))

convert_to_(network)
convert_to_aucell(network, tf, target)
#> $FOXO4
#>  [1] "BCL2L11"  "BCL6"     "CDKN1A"   "CDKN1B"   "G6PC"     "GADD45A" 
#>  [7] "IGFBP1"   "PCK2"     "PPARGC1A" "RBL2"     "SCP2"     "SOD2"    
#> [13] "SPRY2"    "TNFSF10" 
#> 
#> $NFIC
#>  [1] "ALAS1"   "C4BPB"   "COL18A1" "COL1A1"  "CSH1"    "CYP17A1" "CYP1A1" 
#>  [8] "FABP7"   "GFAP"    "GTF3C1"  "GTF3C2"  "GUCY1B3" "HMGB1"   "HSD11B2"
#> [15] "ITGA5"   "JUN"     "MATN1"   "MYC"     "POLD2"   "POLE2"   "PROC"   
#> [22] "SLC2A4"  "SLC34A2" "SMUG1"   "TFAP2A"  "TFPI2"   "TP53"    "VWF"    
#> 
#> $RFXAP
#>  [1] "HLA-DMA"  "HLA-DMB"  "HLA-DOA"  "HLA-DOB"  "HLA-DPA1" "HLA-DPB1"
#>  [7] "HLA-DQA1" "HLA-DQA2" "HLA-DQB1" "HLA-DRA"  "HLA-DRB1" "HLA-DRB3"
#> [13] "HLA-DRB4" "HLA-DRB5"
#> 
#> $SMAD3
#>  [1] "AHR"      "ANKRD1"   "APOC3"    "BAMBI"    "BGLAP"    "CCND2"   
#>  [7] "CDKN1A"   "COL1A1"   "COL3A1"   "COL6A1"   "COL6A3"   "COL7A1"  
#> [13] "CYP17A1"  "EDN1"     "EPO"      "FOXP3"    "FSTL3"    "GLI2"    
#> [19] "ITGA11"   "JUN"      "MET"      "MMP9"     "MYC"      "POR"     
#> [25] "PTHLH"    "SERPINE1" "SMAD7"    "TERT"     "TGFB1"    "TIMP1"   
#> [31] "TNC"      "VEGFA"    "ZFP36"   
#> 
#> $TFAP2A
#>  [1] "ACHE"     "ADM"      "ADRA1A"   "ADRA2B"   "APP"      "BACE1"   
#>  [7] "BCL2"     "CCNB1"    "CDH1"     "CDKN1A"   "CFTR"     "CGA"     
#> [13] "CRABP2"   "CRX"      "CSH1"     "CYP11A1"  "DBH"      "EGFR"    
#> [19] "ERBB2"    "ETS1"     "F2R"      "FAS"      "FMR1"     "FOXA1"   
#> [25] "GALNT3"   "GAST"     "GPX4"     "GUSB"     "HGF"      "HIP1"    
#> [31] "HMOX1"    "HOXA4"    "HSD17B1"  "HTT"      "IGFBP5"   "KIT"     
#> [37] "KRT5"     "L1CAM"    "MIP"      "MMP2"     "MPG"      "MYC"     
#> [43] "NME3"     "NR3C1"    "OAT"      "PENK"     "PLAUR"    "POLD2"   
#> [49] "PRKCA"    "PTGS2"    "PVR"      "S100B"    "SERPINE1" "SLC19A1" 
#> [55] "SMPD1"    "SNAP91"   "SP1"      "SPRY2"    "TAF7"     "TH"      
#> [61] "TIMP2"    "VEGFA"   
#> 
convert_to_gsva(network, tf, target)
#> $FOXO4
#>  [1] "BCL2L11"  "BCL6"     "CDKN1A"   "CDKN1B"   "G6PC"     "GADD45A" 
#>  [7] "IGFBP1"   "PCK2"     "PPARGC1A" "RBL2"     "SCP2"     "SOD2"    
#> [13] "SPRY2"    "TNFSF10" 
#> 
#> $NFIC
#>  [1] "ALAS1"   "C4BPB"   "COL18A1" "COL1A1"  "CSH1"    "CYP17A1" "CYP1A1" 
#>  [8] "FABP7"   "GFAP"    "GTF3C1"  "GTF3C2"  "GUCY1B3" "HMGB1"   "HSD11B2"
#> [15] "ITGA5"   "JUN"     "MATN1"   "MYC"     "POLD2"   "POLE2"   "PROC"   
#> [22] "SLC2A4"  "SLC34A2" "SMUG1"   "TFAP2A"  "TFPI2"   "TP53"    "VWF"    
#> 
#> $RFXAP
#>  [1] "HLA-DMA"  "HLA-DMB"  "HLA-DOA"  "HLA-DOB"  "HLA-DPA1" "HLA-DPB1"
#>  [7] "HLA-DQA1" "HLA-DQA2" "HLA-DQB1" "HLA-DRA"  "HLA-DRB1" "HLA-DRB3"
#> [13] "HLA-DRB4" "HLA-DRB5"
#> 
#> $SMAD3
#>  [1] "AHR"      "ANKRD1"   "APOC3"    "BAMBI"    "BGLAP"    "CCND2"   
#>  [7] "CDKN1A"   "COL1A1"   "COL3A1"   "COL6A1"   "COL6A3"   "COL7A1"  
#> [13] "CYP17A1"  "EDN1"     "EPO"      "FOXP3"    "FSTL3"    "GLI2"    
#> [19] "ITGA11"   "JUN"      "MET"      "MMP9"     "MYC"      "POR"     
#> [25] "PTHLH"    "SERPINE1" "SMAD7"    "TERT"     "TGFB1"    "TIMP1"   
#> [31] "TNC"      "VEGFA"    "ZFP36"   
#> 
#> $TFAP2A
#>  [1] "ACHE"     "ADM"      "ADRA1A"   "ADRA2B"   "APP"      "BACE1"   
#>  [7] "BCL2"     "CCNB1"    "CDH1"     "CDKN1A"   "CFTR"     "CGA"     
#> [13] "CRABP2"   "CRX"      "CSH1"     "CYP11A1"  "DBH"      "EGFR"    
#> [19] "ERBB2"    "ETS1"     "F2R"      "FAS"      "FMR1"     "FOXA1"   
#> [25] "GALNT3"   "GAST"     "GPX4"     "GUSB"     "HGF"      "HIP1"    
#> [31] "HMOX1"    "HOXA4"    "HSD17B1"  "HTT"      "IGFBP5"   "KIT"     
#> [37] "KRT5"     "L1CAM"    "MIP"      "MMP2"     "MPG"      "MYC"     
#> [43] "NME3"     "NR3C1"    "OAT"      "PENK"     "PLAUR"    "POLD2"   
#> [49] "PRKCA"    "PTGS2"    "PVR"      "S100B"    "SERPINE1" "SLC19A1" 
#> [55] "SMPD1"    "SNAP91"   "SP1"      "SPRY2"    "TAF7"     "TH"      
#> [61] "TIMP2"    "VEGFA"   
#> 
convert_to_wmean(network, tf, target, mor, likelihood)
#> # A tibble: 151 × 4
#>    source target     mor likelihood
#>    <chr>  <chr>    <dbl>      <dbl>
#>  1 FOXO4  BCL2L11      1          1
#>  2 FOXO4  BCL6         1          1
#>  3 FOXO4  CDKN1A       1          1
#>  4 FOXO4  CDKN1B       1          1
#>  5 FOXO4  G6PC         1          1
#>  6 FOXO4  GADD45A      1          1
#>  7 FOXO4  IGFBP1       1          1
#>  8 FOXO4  PCK2         1          1
#>  9 FOXO4  PPARGC1A     1          1
#> 10 FOXO4  RBL2         1          1
#> # … with 141 more rows
convert_to_ora(network, tf, target)
#> $FOXO4
#>  [1] "BCL2L11"  "BCL6"     "CDKN1A"   "CDKN1B"   "G6PC"     "GADD45A" 
#>  [7] "IGFBP1"   "PCK2"     "PPARGC1A" "RBL2"     "SCP2"     "SOD2"    
#> [13] "SPRY2"    "TNFSF10" 
#> 
#> $NFIC
#>  [1] "ALAS1"   "C4BPB"   "COL18A1" "COL1A1"  "CSH1"    "CYP17A1" "CYP1A1" 
#>  [8] "FABP7"   "GFAP"    "GTF3C1"  "GTF3C2"  "GUCY1B3" "HMGB1"   "HSD11B2"
#> [15] "ITGA5"   "JUN"     "MATN1"   "MYC"     "POLD2"   "POLE2"   "PROC"   
#> [22] "SLC2A4"  "SLC34A2" "SMUG1"   "TFAP2A"  "TFPI2"   "TP53"    "VWF"    
#> 
#> $RFXAP
#>  [1] "HLA-DMA"  "HLA-DMB"  "HLA-DOA"  "HLA-DOB"  "HLA-DPA1" "HLA-DPB1"
#>  [7] "HLA-DQA1" "HLA-DQA2" "HLA-DQB1" "HLA-DRA"  "HLA-DRB1" "HLA-DRB3"
#> [13] "HLA-DRB4" "HLA-DRB5"
#> 
#> $SMAD3
#>  [1] "AHR"      "ANKRD1"   "APOC3"    "BAMBI"    "BGLAP"    "CCND2"   
#>  [7] "CDKN1A"   "COL1A1"   "COL3A1"   "COL6A1"   "COL6A3"   "COL7A1"  
#> [13] "CYP17A1"  "EDN1"     "EPO"      "FOXP3"    "FSTL3"    "GLI2"    
#> [19] "ITGA11"   "JUN"      "MET"      "MMP9"     "MYC"      "POR"     
#> [25] "PTHLH"    "SERPINE1" "SMAD7"    "TERT"     "TGFB1"    "TIMP1"   
#> [31] "TNC"      "VEGFA"    "ZFP36"   
#> 
#> $TFAP2A
#>  [1] "ACHE"     "ADM"      "ADRA1A"   "ADRA2B"   "APP"      "BACE1"   
#>  [7] "BCL2"     "CCNB1"    "CDH1"     "CDKN1A"   "CFTR"     "CGA"     
#> [13] "CRABP2"   "CRX"      "CSH1"     "CYP11A1"  "DBH"      "EGFR"    
#> [19] "ERBB2"    "ETS1"     "F2R"      "FAS"      "FMR1"     "FOXA1"   
#> [25] "GALNT3"   "GAST"     "GPX4"     "GUSB"     "HGF"      "HIP1"    
#> [31] "HMOX1"    "HOXA4"    "HSD17B1"  "HTT"      "IGFBP5"   "KIT"     
#> [37] "KRT5"     "L1CAM"    "MIP"      "MMP2"     "MPG"      "MYC"     
#> [43] "NME3"     "NR3C1"    "OAT"      "PENK"     "PLAUR"    "POLD2"   
#> [49] "PRKCA"    "PTGS2"    "PVR"      "S100B"    "SERPINE1" "SLC19A1" 
#> [55] "SMPD1"    "SNAP91"   "SP1"      "SPRY2"    "TAF7"     "TH"      
#> [61] "TIMP2"    "VEGFA"   
#> 
convert_to_wsum(network, tf, target, mor)
#> # A tibble: 151 × 4
#>    source target     mor likelihood
#>    <chr>  <chr>    <dbl>      <dbl>
#>  1 FOXO4  BCL2L11      1          1
#>  2 FOXO4  BCL6         1          1
#>  3 FOXO4  CDKN1A       1          1
#>  4 FOXO4  CDKN1B       1          1
#>  5 FOXO4  G6PC         1          1
#>  6 FOXO4  GADD45A      1          1
#>  7 FOXO4  IGFBP1       1          1
#>  8 FOXO4  PCK2         1          1
#>  9 FOXO4  PPARGC1A     1          1
#> 10 FOXO4  RBL2         1          1
#> # … with 141 more rows
convert_to_ulm(network, tf, target, mor)
#> # A tibble: 151 × 4
#>    source target     mor likelihood
#>    <chr>  <chr>    <dbl>      <dbl>
#>  1 FOXO4  BCL2L11      1          1
#>  2 FOXO4  BCL6         1          1
#>  3 FOXO4  CDKN1A       1          1
#>  4 FOXO4  CDKN1B       1          1
#>  5 FOXO4  G6PC         1          1
#>  6 FOXO4  GADD45A      1          1
#>  7 FOXO4  IGFBP1       1          1
#>  8 FOXO4  PCK2         1          1
#>  9 FOXO4  PPARGC1A     1          1
#> 10 FOXO4  RBL2         1          1
#> # … with 141 more rows
convert_to_mlm(network, tf, target, mor)
#> # A tibble: 151 × 4
#>    source target     mor likelihood
#>    <chr>  <chr>    <dbl>      <dbl>
#>  1 FOXO4  BCL2L11      1          1
#>  2 FOXO4  BCL6         1          1
#>  3 FOXO4  CDKN1A       1          1
#>  4 FOXO4  CDKN1B       1          1
#>  5 FOXO4  G6PC         1          1
#>  6 FOXO4  GADD45A      1          1
#>  7 FOXO4  IGFBP1       1          1
#>  8 FOXO4  PCK2         1          1
#>  9 FOXO4  PPARGC1A     1          1
#> 10 FOXO4  RBL2         1          1
#> # … with 141 more rows
convert_to_viper(network, tf, target, mor, likelihood)
#> $FOXO4
#> $FOXO4$tfmode
#>  BCL2L11     BCL6   CDKN1A   CDKN1B     G6PC  GADD45A   IGFBP1     PCK2 
#>        1        1        1        1        1        1        1        1 
#> PPARGC1A     RBL2     SCP2     SOD2    SPRY2  TNFSF10 
#>        1        1        1        1        1        1 
#> 
#> $FOXO4$likelihood
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> 
#> 
#> $NFIC
#> $NFIC$tfmode
#>   ALAS1   C4BPB COL18A1  COL1A1    CSH1 CYP17A1  CYP1A1   FABP7    GFAP  GTF3C1 
#>       1       1       1       1       1       1      -1       1       1       1 
#>  GTF3C2 GUCY1B3   HMGB1 HSD11B2   ITGA5     JUN   MATN1     MYC   POLD2   POLE2 
#>       1       1       1       1      -1       1      -1       1       1       1 
#>    PROC  SLC2A4 SLC34A2   SMUG1  TFAP2A   TFPI2    TP53     VWF 
#>       1       1       1       1       1       1       1       1 
#> 
#> $NFIC$likelihood
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> 
#> 
#> $RFXAP
#> $RFXAP$tfmode
#>  HLA-DMA  HLA-DMB  HLA-DOA  HLA-DOB HLA-DPA1 HLA-DPB1 HLA-DQA1 HLA-DQA2 
#>        1        1        1        1        1        1        1        1 
#> HLA-DQB1  HLA-DRA HLA-DRB1 HLA-DRB3 HLA-DRB4 HLA-DRB5 
#>        1        1        1        1        1        1 
#> 
#> $RFXAP$likelihood
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> 
#> 
#> $SMAD3
#> $SMAD3$tfmode
#>      AHR   ANKRD1    APOC3    BAMBI    BGLAP    CCND2   CDKN1A   COL1A1 
#>        1        1        1        1        1        1        1        1 
#>   COL3A1   COL6A1   COL6A3   COL7A1  CYP17A1     EDN1      EPO    FOXP3 
#>        1        1        1        1       -1        1        1        1 
#>    FSTL3     GLI2   ITGA11      JUN      MET     MMP9      MYC      POR 
#>        1        1        1        1        1        1        1        1 
#>    PTHLH SERPINE1    SMAD7     TERT    TGFB1    TIMP1      TNC    VEGFA 
#>        1        1        1       -1       -1        1        1        1 
#>    ZFP36 
#>        1 
#> 
#> $SMAD3$likelihood
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> 
#> 
#> $TFAP2A
#> $TFAP2A$tfmode
#>     ACHE      ADM   ADRA1A   ADRA2B      APP    BACE1     BCL2    CCNB1 
#>       -1        1        1        1        1        1       -1        1 
#>     CDH1   CDKN1A     CFTR      CGA   CRABP2      CRX     CSH1  CYP11A1 
#>        1        1        1        1        1        1        1        1 
#>      DBH     EGFR    ERBB2     ETS1      F2R      FAS     FMR1    FOXA1 
#>        1        1        1        1        1        1        1        1 
#>   GALNT3     GAST     GPX4     GUSB      HGF     HIP1    HMOX1    HOXA4 
#>        1        1        1        1       -1        1        1       -1 
#>  HSD17B1      HTT   IGFBP5      KIT     KRT5    L1CAM      MIP     MMP2 
#>       -1        1        1        1        1        1        1        1 
#>      MPG      MYC     NME3    NR3C1      OAT     PENK    PLAUR    POLD2 
#>        1        1        1        1        1        1        1        1 
#>    PRKCA    PTGS2      PVR    S100B SERPINE1  SLC19A1    SMPD1   SNAP91 
#>        1        1        1        1        1        1        1        1 
#>      SP1    SPRY2     TAF7       TH    TIMP2    VEGFA 
#>        1        1        1        1        1        1 
#> 
#> $TFAP2A$likelihood
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [39] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> 
#>