decouple

Calculate the TF activity per sample out of a gene expression matrix by coupling a regulon network with a variety of statistics.

decouple(
  mat,
  network,
  .source,
  .target,
  statistics,
  args = list(NULL),
  show_toy_call = FALSE
)

Arguments

mat
Matrix to evaluate (e.g. expression matrix). Target nodes in rows and conditions in columns. rownames(mat) must have at least one intersection with the elements in network .target column.

network
Tibble or dataframe with edges and it's associated metadata.

.source
Column with source nodes.

.target
Column with target nodes.

statistics
Statistical methods to be coupled.

args
A list of argument-lists the same length as statistics (or length 1). The default argument, list(NULL), will be recycled to the same length as statistics, and will call each function with no arguments (apart from mat, network, .source and, .target).

show_toy_call
The call of each statistic must be informed?

Value

A long format tibble of the enrichment scores for each tf across the samples. Resulting tibble contains the following columns:

statistic: Indicates which method is associated with which score.

tf: Source nodes of network.

condition: Condition representing each column of mat.

score: Regulatory activity (enrichment score).

statistic_time: Internal execution time indicator.

...: Columns of metadata generated by certain statistics.

See also

Examples

inputs_dir <- system.file("testdata", "inputs", package = "decoupleR")

mat <- readRDS(file.path(inputs_dir, "input-expr_matrix.rds"))
network <- readRDS(file.path(inputs_dir, "input-dorothea_genesets.rds"))

decouple(
    mat = mat,
    network = network,
    .source = "tf",
    .target = "target",
    statistics = c("gsva", "mean", "pscira", "scira", "viper"),
    args = list(
        gsva = list(verbose = FALSE),
        mean = list(.mor = "mor", .likelihood = "likelihood"),
        pscira = list(.mor = "mor"),
        scira = list(.mor = "mor"),
        viper = list(
            .mor = "mor",
            .likelihood = "likelihood",
            verbose = FALSE
       )
    )
)

#> # A tibble: 112 x 7
#>    run_id statistic tf    condition    score statistic_time p_value
#>    <chr>  <chr>     <chr> <chr>        <dbl> <drtn>           <dbl>
#>  1 1      gsva      FOXO4 GSM2753335 -0.380  7.132416 secs       NA
#>  2 1      gsva      FOXO4 GSM2753336 -0.300  7.132416 secs       NA
#>  3 1      gsva      FOXO4 GSM2753337  0.239  7.132416 secs       NA
#>  4 1      gsva      FOXO4 GSM2753338  0.0907 7.132416 secs       NA
#>  5 1      gsva      NFIC  GSM2753335 -0.0845 7.132416 secs       NA
#>  6 1      gsva      NFIC  GSM2753336  0.0778 7.132416 secs       NA
#>  7 1      gsva      NFIC  GSM2753337 -0.260  7.132416 secs       NA
#>  8 1      gsva      NFIC  GSM2753338  0.281  7.132416 secs       NA
#>  9 1      gsva      RFXAP GSM2753335 -0.810  7.132416 secs       NA
#> 10 1      gsva      RFXAP GSM2753336 -0.472  7.132416 secs       NA
#> # … with 102 more rows