decoupleR: decoupleR: Ensemble of computational methods to infer biological activities from omics data

Many methods allow us to extract biological activities from omics data using information from prior knowledge resources, reducing the dimensionality for increased statistical power and better interpretability. Here, we present decoupleR, a Bioconductor package containing different statistical methods to extract these signatures within a unified framework. decoupleR allows the user to flexibly test any method with any resource. It incorporates methods that take into account the sign and weight of network interactions. decoupleR can be used with any omic, as long as its features can be linked to a biological process based on prior knowledge. For example, in transcriptomics gene sets regulated by a transcription factor, or in phospho-proteomics phosphosites that are targeted by a kinase.

See also

Author

Maintainer: Pau Badia-i-Mompel pau.badia@uni-heidelberg.de (ORCID)

Authors:

Jesús Vélez-Santiago jvelezmagic@gmail.com (ORCID)

Jana Braunger jana.bc@gmx.de (ORCID)

Celina Geiss celina.geiss@stud.uni-heidelberg.de (ORCID)

Daniel Dimitrov daniel.dimitrov@uni-heidelberg.de (ORCID)

Sophia Müller-Dott sophia.mueller-dott@uni-heidelberg.de (ORCID)

Petr Taus petr.taus@ceitec.muni.cz (ORCID)

Aurélien Dugourd aurelien.dugourd@bioquant.uni-heidelberg.de (ORCID)

Christian H. Holland cholland2408@gmail.com (ORCID)

Ricardo O. Ramirez Flores roramirezf@uni-heidelberg.de (ORCID)

Julio Saez-Rodriguez pub.saez@uni-heidelberg.de (ORCID)