SCIRA (Single Cell Inference of Regulatory Activity)

Calculates TF activity according to SCIRA.

run_scira(
  mat,
  network,
  .source = .data$tf,
  .target = .data$target,
  .mor = .data$mor,
  sparse = FALSE,
  fast = TRUE
)

Arguments

mat
Matrix to evaluate (e.g. expression matrix). Target nodes in rows and conditions in columns. rownames(mat) must have at least one intersection with the elements in network .target column.

network
Tibble or dataframe with edges and it's associated metadata.

.source
Column with source nodes.

.target
Column with target nodes.

.mor
Column with edge mode of regulation (i.e. mor).

sparse
Logical value indicating if the generated profile matrix should be sparse.

fast
Logical value indicating if the lineal model must be calculated with speedglm::speedlm.fit() or with base stats::lm().

Value

A long format tibble of the enrichment scores for each tf across the conditions. Resulting tibble contains the following columns:

statistic: Indicates which method is associated with which score.

tf: Source nodes of network.

condition: Condition representing each column of mat.

score: Regulatory activity (enrichment score).

statistic_time: Internal execution time indicator.

Details

Estimation of regulatory activity: A linear regression of the expression profile is performed against the "target profile" of the given TF, where in the target profile, any member of the regulon will be assigned a +1 for activated actions, a -1 for inhibitory activations and, finally, a 0 for all genes that are not a member of the TF regulon. TF activity is then defined as the t-statistic of this linear regression.

See also

Examples

inputs_dir <- system.file("testdata", "inputs", package = "decoupleR")

mat <- readRDS(file.path(inputs_dir, "input-expr_matrix.rds"))
network <- readRDS(file.path(inputs_dir, "input-dorothea_genesets.rds"))

run_scira(mat, network, tf, target, mor)

#> # A tibble: 20 x 5
#>    statistic tf     condition  score statistic_time
#>    <chr>     <chr>  <chr>      <dbl> <drtn>        
#>  1 scira     FOXO4  GSM2753335 2.06  0.1325109 secs
#>  2 scira     FOXO4  GSM2753336 2.05  0.1325109 secs
#>  3 scira     FOXO4  GSM2753337 2.32  0.1325109 secs
#>  4 scira     FOXO4  GSM2753338 2.31  0.1325109 secs
#>  5 scira     NFIC   GSM2753335 0.903 0.1325109 secs
#>  6 scira     NFIC   GSM2753336 0.978 0.1325109 secs
#>  7 scira     NFIC   GSM2753337 0.679 0.1325109 secs
#>  8 scira     NFIC   GSM2753338 0.748 0.1325109 secs
#>  9 scira     SMAD3  GSM2753335 0.548 0.1325109 secs
#> 10 scira     SMAD3  GSM2753336 0.786 0.1325109 secs
#> 11 scira     SMAD3  GSM2753337 1.17  0.1325109 secs
#> 12 scira     SMAD3  GSM2753338 1.20  0.1325109 secs
#> 13 scira     TFAP2A GSM2753335 1.24  0.1325109 secs
#> 14 scira     TFAP2A GSM2753336 1.55  0.1325109 secs
#> 15 scira     TFAP2A GSM2753337 1.75  0.1325109 secs
#> 16 scira     TFAP2A GSM2753338 1.85  0.1325109 secs
#> 17 scira     RFXAP  GSM2753335 1.35  0.1325109 secs
#> 18 scira     RFXAP  GSM2753336 2.11  0.1325109 secs
#> 19 scira     RFXAP  GSM2753337 2.73  0.1325109 secs
#> 20 scira     RFXAP  GSM2753338 2.90  0.1325109 secs