Univariate Linear Model (ULM)

Calculates regulatory activities by fitting univariate linear models (ULM).

run_ulm(
  mat,
  network,
  .source = .data$source,
  .target = .data$target,
  .mor = .data$mor,
  .likelihood = .data$likelihood,
  sparse = FALSE,
  center = FALSE,
  na.rm = FALSE
)

Arguments

mat
Matrix to evaluate (e.g. expression matrix). Target nodes in rows and conditions in columns. rownames(mat) must have at least one intersection with the elements in network .target column.

network
Tibble or dataframe with edges and it's associated metadata.

.source
Column with source nodes.

.target
Column with target nodes.

.mor
Column with edge mode of regulation (i.e. mor).

.likelihood
Column with edge likelihood.

sparse
Deprecated parameter.

center
Logical value indicating if mat must be centered by base::rowMeans().

na.rm
Should missing values (including NaN) be omitted from the calculations of base::rowMeans()?

Value

A long format tibble of the enrichment scores for each source across the samples. Resulting tibble contains the following columns:

statistic: Indicates which method is associated with which score.

source: Source nodes of network.

condition: Condition representing each column of mat.

score: Regulatory activity (enrichment score).

Details

ULM fits a (univariate) linear model to estimate regulatory activities. ULM fits a linear model that predicts the observed molecular readouts using the given weights of a regulator as a single co-variate. The obtained t-value from the fitted model is the activity of the regulator. This approach was first described in: Improved detection of tumor suppressor events in single-cell RNA-Seq data.

See also

Examples

inputs_dir <- system.file("testdata", "inputs", package = "decoupleR")

mat <- readRDS(file.path(inputs_dir, "input-expr_matrix.rds"))
network <- readRDS(file.path(inputs_dir, "input-dorothea_genesets.rds"))

run_ulm(mat, network, .source='tf')
#> # A tibble: 20 × 5
#>    statistic source condition    score p_value
#>    <chr>     <chr>  <chr>        <dbl>   <dbl>
#>  1 ulm       FOXO4  GSM2753335  1.71    0.0897
#>  2 ulm       FOXO4  GSM2753336  1.50    0.135 
#>  3 ulm       FOXO4  GSM2753337  1.67    0.0968
#>  4 ulm       FOXO4  GSM2753338  1.50    0.135 
#>  5 ulm       NFIC   GSM2753335  0.390   0.697 
#>  6 ulm       NFIC   GSM2753336  0.284   0.777 
#>  7 ulm       NFIC   GSM2753337 -0.274   0.785 
#>  8 ulm       NFIC   GSM2753338 -0.344   0.731 
#>  9 ulm       SMAD3  GSM2753335 -0.102   0.919 
#> 10 ulm       SMAD3  GSM2753336 -0.0533  0.958 
#> 11 ulm       SMAD3  GSM2753337  0.115   0.908 
#> 12 ulm       SMAD3  GSM2753338 -0.0278  0.978 
#> 13 ulm       TFAP2A GSM2753335  0.430   0.668 
#> 14 ulm       TFAP2A GSM2753336  0.466   0.642 
#> 15 ulm       TFAP2A GSM2753337  0.310   0.757 
#> 16 ulm       TFAP2A GSM2753338  0.167   0.867 
#> 17 ulm       RFXAP  GSM2753335  0.917   0.361 
#> 18 ulm       RFXAP  GSM2753336  1.57    0.119 
#> 19 ulm       RFXAP  GSM2753337  2.15    0.0331
#> 20 ulm       RFXAP  GSM2753338  2.18    0.0307