VIPER wrapper

This function is a convenient wrapper for the viper::viper() function.

run_viper(
  mat,
  network,
  .source = .data$tf,
  .target = .data$target,
  .mor = .data$mor,
  .likelihood = .data$likelihood,
  ...
)

Arguments

mat
Matrix to evaluate (e.g. expression matrix). Target nodes in rows and conditions in columns. rownames(mat) must have at least one intersection with the elements in network .target column.

network
Tibble or dataframe with edges and it's associated metadata.

.source
Column with source nodes.

.target
Column with target nodes.

.mor
Column with edge mode of regulation (i.e. mor).

.likelihood
Column with edge likelihood.

...
Arguments passed on to viper::viper

dnull

Numeric matrix for the null model, usually generated by nullTtest

pleiotropy

Logical, whether correction for pleiotropic regulation should be performed

nes

Logical, whether the enrichment score reported should be normalized

method

Character string indicating the method for computing the single samples signature, either scale, rank, mad, ttest or none

bootstraps

Integer indicating the number of bootstraps iterations to perform. Only the scale method is implemented with bootstraps.

minsize

Integer indicating the minimum number of targets allowed per regulon

adaptive.size

Logical, whether the weighting scores should be taken into account for computing the regulon size

eset.filter

Logical, whether the dataset should be limited only to the genes represented in the interactome #' @param mvws Number or vector indicating either the exponent score for the metaViper weights, or the inflection point and trend for the sigmoid function describing the weights in metaViper

pleiotropyArgs

list of 5 numbers for the pleotropy correction indicating: regulators p-value threshold, pleiotropic interaction p-value threshold, minimum number of targets in the overlap between pleiotropic regulators, penalty for the pleiotropic interactions and the method for computing the pleiotropy, either absolute or adaptive

cores

Integer indicating the number of cores to use (only 1 in Windows-based systems)

verbose

Logical, whether progression messages should be printed in the terminal

Value

A long format tibble of the enrichment scores for each tf across the conditions. Resulting tibble contains the following columns:

statistic: Indicates which method is associated with which score.

tf: Source nodes of network.

condition: Condition representing each column of mat.

score: Regulatory activity (enrichment score).

statistic_time: Internal execution time indicator.

See also

Examples

inputs_dir <- system.file("testdata", "inputs", package = "decoupleR")

mat <- readRDS(file.path(inputs_dir, "input-expr_matrix.rds"))
network <- readRDS(file.path(inputs_dir, "input-dorothea_genesets.rds"))

run_viper(mat, network, tf, target, mor, likelihood, verbose = FALSE)

#> # A tibble: 12 x 5
#>    statistic tf     condition    score statistic_time 
#>    <chr>     <chr>  <chr>        <dbl> <drtn>         
#>  1 viper     NFIC   GSM2753335  0.0696 0.02561426 secs
#>  2 viper     NFIC   GSM2753336 -0.0265 0.02561426 secs
#>  3 viper     NFIC   GSM2753337 -0.516  0.02561426 secs
#>  4 viper     NFIC   GSM2753338 -0.543  0.02561426 secs
#>  5 viper     SMAD3  GSM2753335  0.176  0.02561426 secs
#>  6 viper     SMAD3  GSM2753336  0.0426 0.02561426 secs
#>  7 viper     SMAD3  GSM2753337  0.219  0.02561426 secs
#>  8 viper     SMAD3  GSM2753338  0.142  0.02561426 secs
#>  9 viper     TFAP2A GSM2753335  0.722  0.02561426 secs
#> 10 viper     TFAP2A GSM2753336  0.582  0.02561426 secs
#> 11 viper     TFAP2A GSM2753337  0.462  0.02561426 secs
#> 12 viper     TFAP2A GSM2753338  0.330  0.02561426 secs