Virtual Inference of Protein-activity by Enriched Regulon analysis (VIPER)

Calculates regulatory activities using VIPER.

Usage

run_viper(
  mat,
  network,
  .source = .data$source,
  .target = .data$target,
  .mor = .data$mor,
  .likelihood = .data$likelihood,
  verbose = FALSE,
  minsize = 5,
  pleiotropy = TRUE,
  eset.filter = FALSE,
  ...
)

Arguments

mat

Matrix to evaluate (e.g. expression matrix). Target nodes in rows and conditions in columns. rownames(mat) must have at least one intersection with the elements in network .target column.

network

Tibble or dataframe with edges and it's associated metadata.

.source

Column with source nodes.

.target

Column with target nodes.

.mor

Column with edge mode of regulation (i.e. mor).

.likelihood

Deprecated argument. Now it will always be set to 1.

verbose

Logical, whether progression messages should be printed in the terminal.

minsize

Integer indicating the minimum number of targets per source.

pleiotropy

Logical, whether correction for pleiotropic regulation should be performed.

eset.filter

Logical, whether the dataset should be limited only to the genes represented in the interactome.

...

Arguments passed on to viper::viper

dnull

Numeric matrix for the null model, usually generated by nullTtest

nes

Logical, whether the enrichment score reported should be normalized

method

Character string indicating the method for computing the single samples signature, either scale, rank, mad, ttest or none

bootstraps

Integer indicating the number of bootstraps iterations to perform. Only the scale method is implemented with bootstraps.

adaptive.size

Logical, whether the weighting scores should be taken into account for computing the regulon size

pleiotropyArgs

list of 5 numbers for the pleotropy correction indicating: regulators p-value threshold, pleiotropic interaction p-value threshold, minimum number of targets in the overlap between pleiotropic regulators, penalty for the pleiotropic interactions and the method for computing the pleiotropy, either absolute or adaptive

cores

Integer indicating the number of cores to use (only 1 in Windows-based systems)

Value

A long format tibble of the enrichment scores for each source across the samples. Resulting tibble contains the following columns:

statistic: Indicates which method is associated with which score.

source: Source nodes of network.

condition: Condition representing each column of mat.

score: Regulatory activity (enrichment score).

Details

This function is a wrapper for the method viper::viper().

See also

Examples

inputs_dir <- system.file("testdata", "inputs", package = "decoupleR")

mat <- readRDS(file.path(inputs_dir, "input-expr_matrix.rds"))
network <- readRDS(file.path(inputs_dir, "input-dorothea_genesets.rds"))

run_viper(mat, network, .source='tf', verbose = FALSE)
#> # A tibble: 20 × 5
#>    statistic source condition  score p_value
#>    <chr>     <chr>  <chr>      <dbl>   <dbl>
#>  1 viper     FOXO4  GSM2753335 1.96  0.0503 
#>  2 viper     FOXO4  GSM2753336 2.00  0.0459 
#>  3 viper     FOXO4  GSM2753337 2.38  0.0173 
#>  4 viper     FOXO4  GSM2753338 2.41  0.0161 
#>  5 viper     NFIC   GSM2753335 0.740 0.459  
#>  6 viper     NFIC   GSM2753336 0.892 0.372  
#>  7 viper     NFIC   GSM2753337 0.579 0.562  
#>  8 viper     NFIC   GSM2753338 0.738 0.461  
#>  9 viper     RFXAP  GSM2753335 1.14  0.256  
#> 10 viper     RFXAP  GSM2753336 2.25  0.0246 
#> 11 viper     RFXAP  GSM2753337 3.04  0.00237
#> 12 viper     RFXAP  GSM2753338 3.26  0.00113
#> 13 viper     SMAD3  GSM2753335 0.831 0.406  
#> 14 viper     SMAD3  GSM2753336 1.04  0.297  
#> 15 viper     SMAD3  GSM2753337 1.50  0.133  
#> 16 viper     SMAD3  GSM2753338 1.72  0.0846 
#> 17 viper     TFAP2A GSM2753335 1.96  0.0500 
#> 18 viper     TFAP2A GSM2753336 2.27  0.0229 
#> 19 viper     TFAP2A GSM2753337 2.50  0.0125 
#> 20 viper     TFAP2A GSM2753338 2.79  0.00524